update and split acceptance tests
